Add files via upload
This script is created to count the number of deaths given diseases in a given year reported by CDC database.
The disease codes are obtained from ICDO dictionary, so you need to put  the icd_nodot script as well as database file in the same folder. The final file is a .CSV file.