Add files via upload
This program is written based on regular expression to count disease (Panic Disorder in this case) death counts in a given year. It can be improved to make the disease code and year more interactively, but this is just an example that it can be nicely done. The disease code is based on ICDo code characters.